Introduce Venue Repository to solve EF issue of not lazy loading backing fields (#51)
EF only does Lazy Loading when you access a Property, and not when you access
backing fields directly. We have some logic that relies on the backing field being
populated. So we have to use eager loading. So introduced a repository to take care
of the eager loading.
Also had a tidy up in Program.cs an made it use the new VenueRepository.